requirements.txt
discord.py==2.5.2
kubernetes==33.1.0
aiohttp==3.12.15
